+++ CHANGES.md
@@ -3,6 +3,9 @@ Unreleased
 
 - Remove `of_fmt` constructor. (#17, @Alizter)
 
+- Add `?bullet` optional argument to `Pp.enumerate` for customizing the bullet.
+  (#TODO, @Alizter)
+
 1.2.0
 -----